Add files via upload
Traffic is a major challenge in many cities around the world today. The number of vehicles is
gradually increasing day by day all over the world. In many instances, traffic flow is
directed by devices such as traffic lights. Traffic lights have become an integral part of
human day-to-day life activities.
This system was designed to control the movement of vehicles in a four-way junction. The
system is designed to allow vehicles from only one way to pass the signal at a time. So that,
the vehicles can direct to any path and unnecessary collisions can be avoided